Removed flx-ido and swiper.
flx-ido has been annoying me since day one, specially because when in a rush I would normally select
the wrong candidate due to the fuzzy matching, I like fuzzy matching as a concept but not for
find-file.

swiper has come and gone, I like the narrowing, but at the same time I don't get used to how it
works.